requirements.txt
global-logger
PyYAML
SQLAlchemy